+++ core/src/lib.rs
@@ -39,7 +39,6 @@ pub mod replay_stage;
 pub mod resource_limits;
 mod result;
 pub mod sample_performance_service;
-#[allow(dead_code)]
 #[cfg(unix)]
 mod scheduler_bindings_server;
 mod shred_fetch_stage;